@@ -2,8 +2,7 @@
   "name": "course-hunt",
   "productName": "Course Hunt",
   "description": "Download the free courses from coursehunters.net easily",
-  "version": "0.0.4",
-  "private": true,
+  "version": "0.0.5",
   "author": "Idrees Dargahwala",
   "main": "app/background.js",
   "build": {
@@ -16,13 +15,12 @@
     "directories": {
       "buildResources": "resources"
     },
-    "publish": null
+    "publish": "null"
   },
   "scripts": {
     "postinstall": "electron-builder install-app-deps",
     "start": "node build/start.js",
-    "publish": "build -p always",
-    "release": "webpack --config=build/webpack.app.config.js --env=production && electron-builder"
+    "release": "webpack --config=build/webpack.app.config.js --env=production && electron-builder -mwl"
   },
   "dependencies": {
     "cheerio": "^1.0.0-rc.2",
@@ -37,13 +35,10 @@
     "@babel/preset-env": "^7.0.0-beta.5",
     "babel-loader": "^8.0.0-beta.0",
     "babel-plugin-transform-object-rest-spread": "^7.0.0-beta.3",
-    "chai": "^4.1.0",
     "css-loader": "^0.28.7",
-    "electron": "1.7.9",
-    "electron-builder": "^19.43.3",
-    "electron-mocha": "^5.0.0",
+    "electron": "^2.0.0",
+    "electron-builder": "^20.13.2",
     "friendly-errors-webpack-plugin": "^1.6.1",
-    "mocha": "^4.0.1",
     "source-map-support": "^0.5.0",
     "spectron": "^3.7.2",
     "style-loader": "^0.19.0",
